Frontend refactoring settings page into modular components
